BugFix: fix the move operation
[Ticket: bugfix]

[body]
1. Fix the bug of moving Line not consistent

[test]
1. pass manual test

+++ src/baseObj.java
@@ -15,7 +15,7 @@ abstract class baseObj implements Serializable {
         this.color = c;
     }
 
-    abstract void translate(int dx, int dy);
+    abstract void translate(double dx, double dy);
 
     abstract void draw(Graphics g);
